requirements.txt
boto3>=1.34.0
pinecone>=5.0.0
pypdf>=4.2.0
python-dotenv>=1.0.1
pytest>=8.0.0
